backend/src/requirements/base.txt
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
django-admin-honeypot==1.1.0
django-admin-honeypot-updated-2021==1.2.0
celery==5.4.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
pytest-factoryboy==2.7.0